Course Detail UI

+++ lib/common/widgets/base_text_widget.dart
@@ -3,7 +3,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 import '../values/colors.dart';
 
-Widget reusableText(String text, {Color color = AppColors.primaryText, int fontSize = 16, FontWeight fontWeight = FontWeight.bold}){
+Widget reusableText(String text, {Color color = AppColors.primaryText, double fontSize = 16, FontWeight fontWeight = FontWeight.bold}){
   return  Text(
     text,
     style: TextStyle(
